feat: adjust arch folder on login page
