+++ src/presentation/components/carrousel/data.ts
@@ -1,17 +1,22 @@
 
+import banner1 from '@/assets/png/bannner1.png'
+import banner2 from '@/assets/png/banner2.png'
+import banner3 from '@/assets/png/banner3.png'
+
+
 export const slides = [
    {
-      image: "/placeholder.svg?height=600&width=800",
+      image: banner1,
       title: "Simplifique a venda de pacotes turísticos",
       description: "Crie e compartilhe links de pagamento personalizados para cada pacote de viagem com rapidez e praticidade.",
    },
    {
-      image: "/placeholder.svg?height=600&width=800",
+      image: banner2,
       title: "Gerencie seus links em tempo real",
       description: "Acompanhe visualizações, status de pagamento e desempenho dos seus pacotes com total controle e transparência.",
    },
    {
-      image: "/placeholder.svg?height=600&width=800",
+      image: banner3,
       title: "Mais segurança e confiança para seus clientes",
       description: "Ofereça uma experiência segura e profissional em cada etapa da jornada de compra, com tecnologia de ponta.",
    },
