+++ src/presentation/components/modals/status-message-modal/data.ts
@@ -49,6 +49,6 @@ export const statusData = [
    {
       status: "IMPORT_REGISTER",
       title: "Usuário importado",
-      description: "Este usuário foi importado por um analista, e ainda não realizou o login.",
+      description: "Este usuário foi importado por um analista, e ainda não realizou o Login.",
    },
 ]
