+++ src/presentation/pages/auth/Login/LoginForm/index.tsx
@@ -1,24 +1,13 @@
 import { Button, Input } from "@/presentation/components/ui"
 import { useState } from "react"
 import { Eye, EyeOff } from "lucide-react"
-import { FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/presentation/components/ui/form"
-import type { LoginProps } from "../types"
-import { Link, useNavigate } from "react-router-dom"
-import { PATHS } from "@/presentation/app/paths"
+import { FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/presentation/components/ui/form.tsx"
+import { Link } from "react-router-dom"
+import type { LoginArgsType } from "@/presentation/pages/auth/Login/LoginForm/types.ts"
 
 
-export function LoginForm({ form, isLoading }: LoginProps) {
+export function LoginForm({ form, isLoading, navigateRegister, navigateForgotPassword }: LoginArgsType) {
    const [showPassword, setShowPassword] = useState(false)
-
-   const navigate = useNavigate()
-
-   const handleNavigateForgotPassword = () => {
-      navigate(PATHS.FORGOT_PASSWORD)
-   }
-
-   const handleNavigateRegister = () => {
-      navigate(PATHS.REGISTER)
-   }
 
    return (
       <div className="flex flex-col gap-6 mt-5 max-w-200 w-full mx-auto">
@@ -49,7 +38,7 @@ export function LoginForm({ form, isLoading }: LoginProps) {
                         <FormLabel>Senha</FormLabel>
                         <a
                            href="#"
-                           onClick={handleNavigateForgotPassword}
+                           onClick={navigateForgotPassword}
                            className="ml-auto text-sm underline-offset-4 hover:underline"
                         >
                            Esqueceu sua senha?
@@ -85,7 +74,7 @@ export function LoginForm({ form, isLoading }: LoginProps) {
                className="w-full bg-primary text-white cursor-pointer"
                disabled={isLoading}
             >
-               {isLoading ? "Entrando..." : "Entraaaaar"}
+               {isLoading ? "Entrando..." : "Entrar"}
             </Button>
 
             <div className="after:border-border relative text-center text-sm after:absolute after:inset-0 after:top-1/2 after:z-0 after:flex after:items-center after:border-t">
@@ -96,7 +85,7 @@ export function LoginForm({ form, isLoading }: LoginProps) {
          </div>
 
          <div className="text-center text-sm">
-            <Link to="#" onClick={handleNavigateRegister} className="underline underline-offset-4">
+            <Link to="#" onClick={navigateRegister} className="underline underline-offset-4">
                Cadastre-se
             </Link>
          </div>
