--- src/presentation/pages/auth/login/types.ts
@@ -1,8 +1,0 @@
-import type { UseFormReturn } from "react-hook-form"
-import type { LoginRequest } from "./schema/login-schema"
-
-
-export interface LoginProps {
-   form: UseFormReturn<LoginRequest>
-   isLoading: boolean
-}
