+++ src/presentation/pages/index.ts
@@ -1,3 +1,3 @@
-export * from "./auth/login"
+export * from "./auth/Login"
 export * from "./auth/reset-password"
 export * from "./auth/forgot-password"
